# [RyuJIT] Devirtualize Comparer<T>.Default