--- a/src/libraries/System.Private.CoreLib/src/System/Collections/Generic/Comparer.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/Collections/Generic/Comparer.cs
@@ -55,6 +55,7 @@ namespace System.Collections.Generic
     // Needs to be public to support binary serialization compatibility
     public sealed partial class GenericComparer<T> : Comparer<T> where T : IComparable<T>
     {
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public override int Compare(T? x, T? y)
         {
             if (x != null)