… src/coreclr/ToolBox/superpmi/superpmi-shared/lwmlist.h
… src/coreclr/inc/jiteeversionguid.h
… src/coreclr/jit/ICorJitInfo_API_names.h
… src/coreclr/tools/Common/JitInterface/CorInfoBase.cs
… src/coreclr/tools/Common/JitInterface/ThunkGenerator/ThunkInput.txt
… src/tests/JIT/opt/Devirtualization/Comparer_get_Default.csproj
… src/tests/issues.targets