See https://github.com/dotnet/runtime/issues/43796

Mostly a copy-paste from the existing `EqualityComparer<T>.Default` devirtualization.
Example:
```csharp
private static int Compare(byte a, byte b) => Comparer<byte>.Default.Compare(a, b);
```
master:
```asm
; Method Program:Compare(ubyte,ubyte):int
G_M13367_IG01:              ;; offset=0000H
       57                   push     rdi
       56                   push     rsi
       4883EC28             sub      rsp, 40
       8BF1                 mov      esi, ecx
       8BFA                 mov      edi, edx
G_M13367_IG02:              ;; offset=000AH
       48B928002535FF7F0000 mov      rcx, 0x7FFF35250028
       BA08000000           mov      edx, 8
       E8F218755F           call     CORINFO_HELP_CLASSINIT_SHARED_DYNAMICCLASS
       48B9D02C0098F8010000 mov      rcx, 0x1F898002CD0
       488B09               mov      rcx, gword ptr [rcx]
       400FB6D6             movzx    rdx, sil
       440FB6C7             movzx    r8, dil
       488B01               mov      rax, qword ptr [rcx]
       488B4048             mov      rax, qword ptr [rax+72]
       488B4020             mov      rax, qword ptr [rax+32]
G_M13367_IG03:              ;; offset=003EH
       4883C428             add      rsp, 40
       5E                   pop      rsi
       5F                   pop      rdi
       48FFE0               rex.jmp  rax
; Total bytes of code: 71
```
PR:
```asm
; Method Program:Compare(ubyte,ubyte):int
G_M13367_IG01:              ;; offset=0000H
G_M13367_IG02:              ;; offset=0000H
       0FB6C1               movzx    rax, cl
       0FB6D2               movzx    rdx, dl
       2BC2                 sub      eax, edx
G_M13367_IG03:              ;; offset=0008H
       C3                   ret      
; Total bytes of code: 9

```
JIT diff is empty but it should allow to write generic versions of `Max<T>`/`Min<T>` functions (https://github.com/dotnet/runtime/issues/43796).