@EgorBo 2021-02-11
> Seems like the enum comparer gets inlined and optimized w/o needing annotation, but can you double-check?
> 
> Did you try crossgenning your test and then running it? Would be good to get somebody more familiar with crossgen2 to sign off on the changes there.
> 
> Changes LGTM, just some comments on comments.

Yeah, the tests pass with crossgen2