@AndyAyersMS 2021-02-11
@dotnet/crossgen-contrib can somebody take a quick look at the crossgen2 changes here?