@eiriktsarpalis 2021-02-11
I ran my [PriorityQueue benchmarks](https://github.com/dotnet/runtime/pull/46009#discussion_r558559981) against the PR branch: this compares a heap implementation using `Comparer<T>.Default.Compare` calls versus an identical one using `IComparable`:

## master branch

|                        Method |    Size |             Mean |           Error |           StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |        
|------------------------------ |-------- |-----------------:|----------------:|-----------------:|------:|------:|------:|----------:|
| PriorityQueue_ComparerDefault |      10 |         502.7 ns |         9.83 ns |         13.13 ns |     - |     - |     - |         - |        
|      PriorityQueue_Comparable |      10 |         353.8 ns |         6.86 ns |          7.90 ns |     - |     - |     - |         - |
| PriorityQueue_ComparerDefault |      50 |       4,171.1 ns |        76.74 ns |         68.02 ns |     - |     - |     - |         - |
|      PriorityQueue_Comparable |      50 |       2,380.8 ns |        37.08 ns |         32.87 ns |     - |     - |     - |         - |        
| PriorityQueue_ComparerDefault |     150 |      20,487.9 ns |       364.63 ns |        341.08 ns |     - |     - |     - |         - |
|      PriorityQueue_Comparable |     150 |      11,740.0 ns |       138.82 ns |        123.06 ns |     - |     - |     - |         - |        
| PriorityQueue_ComparerDefault |     500 |      92,828.3 ns |     1,767.72 ns |      1,567.04 ns |     - |     - |     - |         - |        
|      PriorityQueue_Comparable |     500 |      64,291.4 ns |     1,196.99 ns |      1,378.46 ns |     - |     - |     - |         - |
| PriorityQueue_ComparerDefault |    1000 |     205,028.1 ns |     3,697.23 ns |      4,257.74 ns |     - |     - |     - |         - |
|      PriorityQueue_Comparable |    1000 |     147,682.6 ns |     2,176.21 ns |      1,929.15 ns |     - |     - |     - |         - |        
| PriorityQueue_ComparerDefault |   10000 |   2,718,728.5 ns |    54,288.34 ns |     58,087.92 ns |     - |     - |     - |       1 B |        
|      PriorityQueue_Comparable |   10000 |   1,935,928.9 ns |    36,901.95 ns |     34,518.11 ns |     - |     - |     - |       1 B |
| PriorityQueue_ComparerDefault | 1000000 | 454,903,575.0 ns | 8,828,957.20 ns | 11,480,138.23 ns |     - |     - |     - |     144 B |
|      PriorityQueue_Comparable | 1000000 | 348,606,720.0 ns | 6,908,734.31 ns |  6,462,434.36 ns |     - |     - |     - |    2360 B |

## PR branch

|                       Method  |    Size |             Mean |           Error |          StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------------ |-------- |-----------------:|----------------:|----------------:|------:|------:|------:|----------:|
| PriorityQueue_ComparerDefault |      10 |         349.2 ns |         6.19 ns |         7.83 ns |     - |     - |     - |         - |
|      PriorityQueue_Comparable |      10 |         327.2 ns |         6.28 ns |         5.87 ns |     - |     - |     - |         - |
| PriorityQueue_ComparerDefault |      50 |       2,514.8 ns |        18.71 ns |        14.61 ns |     - |     - |     - |         - |
|      PriorityQueue_Comparable |      50 |       2,274.4 ns |        37.11 ns |        32.89 ns |     - |     - |     - |         - |
| PriorityQueue_ComparerDefault |     150 |      12,616.7 ns |        54.33 ns |        45.37 ns |     - |     - |     - |         - |
|      PriorityQueue_Comparable |     150 |      11,299.3 ns |        96.31 ns |        90.09 ns |     - |     - |     - |         - |
| PriorityQueue_ComparerDefault |     500 |      66,912.2 ns |       541.71 ns |       506.72 ns |     - |     - |     - |         - |
|      PriorityQueue_Comparable |     500 |      63,575.9 ns |       375.87 ns |       333.20 ns |     - |     - |     - |         - |
| PriorityQueue_ComparerDefault |    1000 |     152,907.1 ns |       554.02 ns |       518.23 ns |     - |     - |     - |         - |
|      PriorityQueue_Comparable |    1000 |     147,177.2 ns |     1,111.73 ns |     1,039.92 ns |     - |     - |     - |         - |
| PriorityQueue_ComparerDefault |   10000 |   2,060,862.3 ns |    33,190.78 ns |    29,422.76 ns |     - |     - |     - |       1 B |
|      PriorityQueue_Comparable |   10000 |   1,965,759.2 ns |    39,021.23 ns |    53,412.70 ns |     - |     - |     - |       1 B |
| PriorityQueue_ComparerDefault | 1000000 | 372,963,264.3 ns | 7,253,139.17 ns | 6,429,718.95 ns |     - |     - |     - |     760 B |
|      PriorityQueue_Comparable | 1000000 | 362,177,334.8 ns | 7,176,855.04 ns | 9,076,423.69 ns |     - |     - |     - |     144 B |