@AndyAyersMS 2021-02-12
Nice work @EgorBo!

> I ran my PriorityQueue benchmarks ...

@adamsitnik Would be nice to fix the benchmark harness so you can pass multiple --corerun args and have it compare them, like it can do for --runtimes.