@stephentoub 2021-02-12
> Would be nice to fix the benchmark harness so you can pass multiple --corerun args and have it compare them, like it can do for --runtimes.

You can, unless I'm misunderstanding what you're asking for.  I do this all the time:
```
--corerun d:\coreclrtest\master\corerun.exe d:\coreclrtest\pr\corerun.exe
```