@AndyAyersMS 2021-02-12
> I do this all the time ...

Guess I'm confused. Maybe what I was thinking was allowing a mixture of --runtimes and --corerun .