@akoeplinger 2021-02-23
@EgorBo this seems to cause a warning in CI, would you mind fixing it:

```
/__w/1/s/src/tests/JIT/opt/Devirtualization/Comparer_get_Default.cs(240,32): warning CS8632: The annotation for nullable reference types should only be used in code within a '#nullable' annotations context. [/__w/1/s/src/tests/JIT/opt/Devirtualization/Comparer_get_Default.csproj]
```